+++ CHANGELOG.md
@@ -30,6 +30,12 @@ We use the following categories for changes:
 
 ### Security
 
+## [0.3.15] 2022-11-11
+
+### Fixed
+
+- merge values override recursive [#sc-5644]
+
 ## [0.3.14] 2022-10-20
 
 ### Added